+++ lib/chef/knife/ec2_server_list.rb
@@ -17,7 +17,7 @@
 # limitations under the License.
 #
 
-require_relative "ec2_base"
+require_relative "helpers/ec2_base"
 
 class Chef
   class Knife
