+++ lib/chef/knife/ec2_vpc_list.rb
@@ -16,7 +16,7 @@
 # limitations under the License.
 #
 
-require_relative "ec2_base"
+require_relative "helpers/ec2_base"
 
 class Chef
   class Knife
