+++ lib/chef/knife/helpers/ec2_base.rb
