+++ lib/chef/knife/helpers/s3_source.rb
